# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CommandExecution": {
        "properties": {
          "approved_command": {
            "title": "Approved Command",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "output_limit_bytes": {
            "title": "Output Limit Bytes",
            "type": "integer"
          },
          "output_truncated": {
            "default": false,
            "title": "Output Truncated",
            "type": "boolean"
          },
          "proposed_step_id": {
            "title": "Proposed Step Id",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "sanitized_stderr": {
            "default": "",
            "title": "Sanitized Stderr",
            "type": "string"
          },
          "sanitized_stdout": {
            "default": "",
            "title": "Sanitized Stdout",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/CommandExecutionStatus"
          },
          "target_host": {
            "title": "Target Host",
            "type": "string"
          },
          "target_port": {
            "title": "Target Port",
            "type": "integer"
          },
          "target_username": {
            "title": "Target Username",
            "type": "string"
          },
          "timeout_s": {
            "title": "Timeout S",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "run_id",
          "proposed_step_id",
          "approved_command",
          "status",
          "target_host",
          "target_port",
          "target_username",
          "timeout_s",
          "output_limit_bytes"
        ],
        "title": "CommandExecution",
        "type": "object"
      },
      "CommandExecutionStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "timed_out",
          "aborted"
        ],
        "title": "CommandExecutionStatus",
        "type": "string"
      },
      "CommandOutputChunk": {
        "properties": {
          "command_execution_id": {
            "title": "Command Execution Id",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "redacted": {
            "default": false,
            "title": "Redacted",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "stream": {
            "title": "Stream",
            "type": "string"
          }
        },
        "required": [
          "id",
          "command_execution_id",
          "run_id",
          "sequence",
          "stream",
          "content",
          "created_at"
        ],
        "title": "CommandOutputChunk",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManualStepCreate": {
        "properties": {
          "command": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Command",
            "type": "string"
          },
          "entered_by": {
            "default": "technician",
            "maxLength": 120,
            "minLength": 1,
            "title": "Entered By",
            "type": "string"
          },
          "expected_signal": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Signal"
          },
          "phase": {
            "default": "diagnostic",
            "maxLength": 80,
            "minLength": 1,
            "title": "Phase",
            "type": "string"
          },
          "purpose": {
            "default": "Manual technician command.",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Purpose",
            "type": "string"
          },
          "timeout_s": {
            "anyOf": [
              {
                "maximum": 600.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout S"
          }
        },
        "required": [
          "command"
        ],
        "title": "ManualStepCreate",
        "type": "object"
      },
...
      "OutboxStatus": {
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed",
          "dead_letter"
        ],
        "title": "OutboxStatus",
        "type": "string"
      },
      "ProposedStep": {
        "properties": {
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "approved_command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Command"
          },
          "command": {
            "title": "Command",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expected_signal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Signal"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "risk_class": {
            "title": "Risk Class",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "safety_notes": {
            "items": {
              "type": "string"
            },
            "title": "Safety Notes",
            "type": "array"
          },
          "safety_summary": {
            "title": "Safety Summary",
            "type": "string"
          },
          "safety_verdict": {
            "title": "Safety Verdict",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/StepStatus"
          },
          "timeout_s": {
            "title": "Timeout S",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "run_id",
          "created_at",
          "source",
          "phase",
          "command",
          "purpose",
          "risk_class",
          "safety_verdict",
          "safety_summary",
          "status",
          "timeout_s"
        ],
        "title": "ProposedStep",
        "type": "object"
      },
...
      "RunEvent": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "approval_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Status"
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "risk_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Class"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "sanitized_stderr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sanitized Stderr"
          },
          "sanitized_stdout": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sanitized Stdout"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "created_at",
          "actor",
          "event_type",
          "summary"
        ],
        "title": "RunEvent",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend control plane for Phoenix ERP, technician approvals, SSH execution, and audit state.",
    "title": "techbold AI Service Desk Autopilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs/{run_id}/commands": {
      "get": {
        "operationId": "get_run_commands_api_runs__run_id__commands_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommandExecution"
                  },
                  "title": "Response Get Run Commands Api Runs  Run Id  Commands Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Commands"
      }
    },
...
    "/api/runs/{run_id}/events": {
      "get": {
        "operationId": "get_run_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunEvent"
                  },
                  "title": "Response Get Run Events Api Runs  Run Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Events"
      }
    },
    "/api/runs/{run_id}/events/stream": {
      "get": {
        "operationId": "stream_run_events_api_runs__run_id__events_stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Run Events"
      }
    },
...
    "/api/runs/{run_id}/manual-step": {
      "post": {
        "operationId": "submit_manual_step_api_runs__run_id__manual_step_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualStepCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposedStep"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Manual Step"
      }
    },
    "/api/runs/{run_id}/manual-steps": {
      "post": {
        "operationId": "submit_manual_step_api_runs__run_id__manual_steps_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualStepCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposedStep"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Manual Step"
      }
    },
    "/api/runs/{run_id}/outbox-events": {
      "get": {
        "operationId": "get_run_outbox_events_api_runs__run_id__outbox_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/OutboxStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutboxEvent"
                  },
                  "title": "Response Get Run Outbox Events Api Runs  Run Id  Outbox Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Outbox Events"
      }
    },
    "/api/runs/{run_id}/outbox-events/dead-letter": {
      "get": {
        "operationId": "get_run_dead_letter_outbox_events_api_runs__run_id__outbox_events_dead_letter_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutboxEvent"
                  },
                  "title": "Response Get Run Dead Letter Outbox Events Api Runs  Run Id  Outbox Events Dead Letter Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Dead Letter Outbox Events"
      }
    },
    "/api/runs/{run_id}/output-chunks": {
      "get": {
        "operationId": "get_run_output_chunks_api_runs__run_id__output_chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommandOutputChunk"
                  },
                  "title": "Response Get Run Output Chunks Api Runs  Run Id  Output Chunks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Output Chunks"
      }
    },
...
    "/api/runs/{run_id}/steps/{step_id}/approve": {
      "post": {
        "operationId": "approve_step_api_runs__run_id__steps__step_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "step_id",
            "required": true,
            "schema": {
              "title": "Step Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandApproval"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Step"
      }
    },
    "/api/runs/{run_id}/steps/{step_id}/edit-and-approve": {
      "post": {
        "operationId": "edit_and_approve_step_api_runs__run_id__steps__step_id__edit_and_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "step_id",
            "required": true,
            "schema": {
              "title": "Step Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandEditApproval"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit And Approve Step"
      }
    },
    "/api/runs/{run_id}/steps/{step_id}/reject": {
      "post": {
        "operationId": "reject_step_api_runs__run_id__steps__step_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "step_id",
            "required": true,
            "schema": {
              "title": "Step Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepRejection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Step"
      }
    },
    "/api/runs/{run_id}/stream": {
      "get": {
        "operationId": "stream_run_events_api_runs__run_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Run Events"
      }
    },
...
  }
}